+++ src/components/comment/Comment.js
@@ -1,20 +1,14 @@
 import React from 'react';
-import { FaPencilAlt, FaTrash } from 'react-icons/fa';
+
+import CommentInput from './CommentInput';
+import CommentList from './CommentList';
 
 const Comment = () => {
   return (
-    <div className="flex justify-between mt-2">
-      <div className="flex">
-        <p className="font-semibold pr-2">닉네임</p>
-        <p className="font-normal">여기에 댓글이 들어갑니다.</p>
-      </div>
-      <div>
-        <button className="mr-2">
-          <FaPencilAlt />
-        </button>
-        <button>
-          <FaTrash />
-        </button>
+    <div>
+      <div className="card w-[34rem] h-[32rem]">
+        <CommentInput />
+        <CommentList />
       </div>
     </div>
   );
